fix duplicate filters type
